Add tests for sendToKafka method
